WIP: Add isEditable boolean for opening each task contents.
