--- frontend/app/src/TaskAway/Task/TaskItem.js
@@ -1,11 +1,0 @@
-// Module Imports
-import React from 'react';
-
-// Relative imports
-import EditTask from '../TaskMethods/EditTask';
-
-export default function TaskItem(props) {
-    return (
-        <EditTask task={props} />
-    );
-}
