+++ frontend/app/src/TaskAway/TaskMethods/EditTask.js
@@ -1,18 +1,12 @@
 // Module Imports
-import React, { useState, useContext } from 'react';
+import React, { useContext } from 'react';
 
 // Relative imports
 import {Context} from '../TaskContextMain';
 import TaskAwayAPI from '../api/api';
-import DeleteTask from './DeleteTask';
 
 export default function EditTasks(props) {
 	const [state, dispatch] = useContext(Context);
-    const [isEditable, setIsEditable] = useState(false);
-
-    const handleIsEditable = () => {
-        setIsEditable(!isEditable);
-    };
 
     const handleIsComplete = () => {
 		dispatch({ type: 'SET_IS_COMPLETE', payload: !state.isComplete })
@@ -20,7 +14,6 @@ export default function EditTasks(props) {
 
     const handleUpdateTodo = () => {
         TaskAwayAPI.updateTask(props.task.task._id, state.title, state.desc, state.isComplete);
-        handleIsEditable();
             
         dispatch(
 			{ 
@@ -41,56 +34,35 @@ export default function EditTasks(props) {
 			}
 		);
     };
-
+    
     return (
         <div className="EditTask">
-            { (!isEditable)
-                ?
-                    <div>
-                        <span style={{ fontWeight: 'bold' }}>
-                            { props.task.task.title }
-                        </span>
-                        { props.task.task.description }
-                        { props.task.task.completed ? "completed" : "not completed" }
-                    </div>
-                :
-                    <div>
-                        <input 
-                            onChange={(event) => dispatch({ type: 'SET_TITLE', payload: event.target.value })}
-                            value={state.title}
-                            placeholder={props.task.task.title}
-                        />
-                        <input 
-                            onChange={event => dispatch({ type: 'SET_DESCRIPTION', payload: event.target.value })}
-                            value={state.desc}
-                            placeholder={props.task.task.description}
-                        />
-                        <label>
-                            Completed
-                            <input 
-                                onClick={handleIsComplete}
-                                value={props.task.task.completed}
-                                type={'checkbox'}
-                            />
-                        </label>
-                        <button 
-                            onClick={() => handleUpdateTodo()}
-                            style={{ color: 'blue'}}
-                        >
-                            Update
-                        </button>
-                    </div>
-                // End-of-Ternary
-            }
-            
-            <button 
-                onClick={handleIsEditable}
-                style={{ color: 'green'}}
-            >
-                Edit
-            </button>
-
-            <DeleteTask taskID={props.task.task._id} />
+            <div>
+                <input 
+                    onChange={(event) => dispatch({ type: 'SET_TITLE', payload: event.target.value })}
+                    value={state.title}
+                    placeholder={props.task.task.title}
+                />
+                <input 
+                    onChange={event => dispatch({ type: 'SET_DESCRIPTION', payload: event.target.value })}
+                    value={state.desc}
+                    placeholder={props.task.task.description}
+                />
+                <label>
+                    Completed
+                    <input 
+                        onClick={handleIsComplete}
+                        value={props.task.task.completed}
+                        type={'checkbox'}
+                    />
+                </label>
+                <button 
+                    onClick={() => handleUpdateTodo()}
+                    style={{ color: 'blue'}}
+                >
+                    Update
+                </button>
+            </div>
         </div>
     );
 }
